requirements.txt
# Core dependencies for the pipeline
openai>=1.0.0
pyyaml>=6.0
rouge-score>=0.1.2
